First approach to user authorization

+++ aquarius/app/assets.py
@@ -10,6 +10,7 @@ from flask import Blueprint, jsonify, request, Response
 from oceandb_driver_interface.search_model import FullTextModel, QueryModel
 from plecos.plecos import (is_valid_dict_local, is_valid_dict_remote, list_errors_dict_local,
                            list_errors_dict_remote)
+from web3 import Web3
 
 from aquarius.app.dao import Dao
 from aquarius.config import Config
@@ -22,6 +23,7 @@ assets = Blueprint('assets', __name__)
 # Prepare OceanDB
 dao = Dao(config_file=app.config['CONFIG_FILE'])
 logger = logging.getLogger('aquarius')
+web3 = Web3('http://localhost:8545')
 
 
 @assets.route('', methods=['GET'])
@@ -460,6 +462,8 @@ def update(did):
     msg, status = validate_date_format(data['created'])
     if msg:
         return msg, status
+    # if not _is_the_creator(dao.get(did), signature=signature):
+    #     return 'You can not update this ddo', 400
     _record = dict()
     _record = copy.deepcopy(data)
     _record['created'] = datetime.strptime(data['created'], '%Y-%m-%dT%H:%M:%SZ')
@@ -512,6 +516,8 @@ def retire(did):
         if dao.get(did) is None:
             return 'This asset DID is not in OceanDB', 404
         else:
+            if not _is_the_creator(dao.get(did), signature=request.headers.get('signature')):
+                return 'You can not delete this ddo', 400
             dao.delete(did)
             return 'Succesfully deleted', 200
     except Exception as err:
@@ -784,3 +790,11 @@ def _reorder_services(services):
         if service['type'] != 'Metadata':
             result.append(service)
     return result
+
+
+def _is_the_creator(ddo, signature):
+    if ddo['proof']['creator'] == web3.personal.ecRecover(
+            message=ddo['id'] + ':' + ddo['proof']['created'],
+            signature=signature):
+        return True
+    return False
